+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/Identity/OnlineAsyncTaskAccelByteSimultaneousLogin.cpp
@@ -23,6 +23,7 @@ FOnlineAsyncTaskAccelByteSimultaneousLogin::FOnlineAsyncTaskAccelByteSimultaneou
 	: FOnlineAsyncTaskAccelByteLogin(InABSubsystem, InLocalUserNum, InAccountCredentials, bInCreateHeadlessAccount)
 {
 	LocalUserNum = INVALID_CONTROLLERID;
+	bByPassSupportedNativeSubsystemCheck = true;
 }
 
 bool FOnlineAsyncTaskAccelByteSimultaneousLogin::IsInitializeAllowed()
@@ -41,7 +42,7 @@ bool FOnlineAsyncTaskAccelByteSimultaneousLogin::IsInitializeAllowed()
 		return false; 
 	}
 
-	if (!Subsystem->IsNativeSubsystemSupported(FName(SecondaryPlatformName)))
+	if (!Subsystem->IsNativeSubsystemSupported(FName(SecondaryPlatformName)) && !bByPassSupportedNativeSubsystemCheck)
 	{
 		AB_OSS_ASYNC_TASK_TRACE_END_VERBOSITY(Warning, TEXT("Failed to do simultaneous login, specified secondary platform is not supported (%s)"), *SecondaryPlatformName);
 		return false;
@@ -188,18 +189,15 @@ void FOnlineAsyncTaskAccelByteSimultaneousLogin::PerformLogin(const FOnlineAccou
 		const THandler<FAccelByteModelsLoginQueueTicketInfo> OnLoginSuccessDelegate = TDelegateUtils<THandler<FAccelByteModelsLoginQueueTicketInfo>>::CreateThreadSafeSelfPtr(this, &FOnlineAsyncTaskAccelByteSimultaneousLogin::OnLoginSuccessV4);
 		const FOAuthErrorHandler OnLoginErrorOAuthDelegate = TDelegateUtils<FOAuthErrorHandler>::CreateThreadSafeSelfPtr(this, &FOnlineAsyncTaskAccelByteSimultaneousLogin::OnLoginErrorOAuth);
 
-		auto NativePlatformLoginType = FOnlineSubsystemAccelByteUtils::GetAccelByteLoginTypeFromNativeSubsystem(FName(NativePlatformCredentials.Type));
-		auto SecondaryPlatformLoginType = FOnlineSubsystemAccelByteUtils::GetAccelByteLoginTypeFromNativeSubsystem(FName(Credentials.Type));
-
 		ApiClient->User.LoginWithSimultaneousPlatformV4(
-			ConvertOSSTypeToAccelBytePlatformType(NativePlatformLoginType),
+			NativePlatformCredentials.Type.ToLower(),
 			NativePlatformTicket,
-			ConvertOSSTypeToAccelBytePlatformType(SecondaryPlatformLoginType),
+			Credentials.Type.ToLower(),
 			SecondaryPlatformTicket,
 			OnLoginSuccessDelegate,
 			OnLoginErrorOAuthDelegate);
 		
-		PlatformId = FAccelByteUtilities::GetUEnumValueAsString(ConvertOSSTypeToAccelBytePlatformType(NativePlatformLoginType));
+		PlatformId = NativePlatformCredentials.Type;
 	}
 }