+++ Source/OnlineSubsystemAccelByte/Private/AsyncTasks/SessionV2/OnlineAsyncTaskAccelByteFindV2GameSessionById.cpp
@@ -78,6 +78,9 @@ void FOnlineAsyncTaskAccelByteFindV2GameSessionById::TriggerDelegates()
 
 	Delegate.ExecuteIfBound(LocalUserNum, bWasSuccessful, FoundSessionResult);
 
+	const TSharedPtr<FOnlineSessionV2AccelByte, ESPMode::ThreadSafe> SessionInterface = StaticCastSharedPtr<FOnlineSessionV2AccelByte>(Subsystem->GetSessionInterface());
+	SessionInterface->SetFindMatchmakingGameSessionByIdInProgress(false);
+
 	AB_OSS_ASYNC_TASK_TRACE_END(TEXT(""));
 }